+++ packages/terra-action-footer/CHANGELOG.md
@@ -6,6 +6,8 @@ Unreleased
 ### Added
 * Added change log to the documentation on terra-dev-site
 
+* Enhanced CenteredActionFooter to have isBlock prop that forces the center socket node to fill the entirety of the footer.
+
 1.13.0 - (June 22, 2018)
 ------------------
 ### Changed